# Speed up setup test by not installing docs